Flarum 2.0 upgrade (#137)
* chore(2.0): code formatting with prettier

Start by formatting your code with Prettier.

* chore(2.0): update dependencies

Update dependencies to Flarum 2.0 compatible versions.

* chore(2.0): update infrastructure

Update the extension infrastructure

* chore(2.0): use new import from extension format

Importing modules from an extension can now be done though `ext:vendor/extension/module-path` format.

* chore(2.0): adapt to extending lazy modules

Some Flarum modules are now lazy loaded. Extending them requires a different approach.

* chore(2.0): misc frontend changes

Miscellaneous frontend changes

* chore(2.0): misc backend changes

Miscellaneous backend changes

* chore(2.0): JSON:API changes

Flarum 2.0 completely changes the JSON:API implementation

* chore(2.0): Backend tests + PHPUnit 9 to 11 changes

Flarum 2.0 uses PHPUnit 11 and encourages use of model factories for easier cross-database testing.

* chore(2.0): `LESS` code changes

Many variables have been renamed to light/dark specific names and most are now used as CSS variables instead.

* chore: Flarum 2.0 upgrade

* Apply fixes from StyleCI

* misc fixes

* more fixes

* more fixes

---------

+++ src/Api/Controllers/DeleteSocialMediaImageController.php
@@ -12,11 +12,10 @@
 namespace FoF\Seo\Api\Controllers;
 
 use Flarum\Api\Controller\AbstractDeleteController;
+use Flarum\Http\RequestUtil;
 use Flarum\Settings\SettingsRepositoryInterface;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Contracts\Filesystem\Cloud;
-use Laminas\Diactoros\Response\EmptyResponse;
-use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 class DeleteSocialMediaImageController extends AbstractDeleteController
@@ -30,18 +29,16 @@ class DeleteSocialMediaImageController extends AbstractDeleteController
         $this->disk = $container->make('filesystem')->disk('flarum-assets');
     }
 
-    protected function delete(ServerRequestInterface $request): ResponseInterface
+    protected function delete(ServerRequestInterface $request): void
     {
-        $request->getAttribute('actor')->assertAdmin();
+        RequestUtil::getActor($request)->assertAdmin();
 
         $path = $this->settings->get('seo_social_media_image_path');
         $this->settings->set('seo_social_media_image_path', null);
         $this->settings->set('seo_social_media_image_url', null);
 
-        if ($this->disk->exists($path)) {
+        if ($path && $this->disk->exists($path)) {
             $this->disk->delete($path);
         }
-
-        return new EmptyResponse(204);
     }
 }
